Split build target field extraction
